+++ modules/board.js
@@ -14,6 +14,7 @@ var Board = function(width, height, arrangement, captures) {
     this.captures = captures ? { '-1': captures['-1'], '1': captures['1'] } : { '-1': 0, '1': 0 }
     this.arrangement = {}
     this.markups = {}
+    this.ghosts = []
     this.lines = []
 
     // Initialize arrangement
@@ -302,8 +303,8 @@ Board.prototype = {
         var done = {}
         var result = []
 
-        for (var i = 0; i < this.width; i++) {
-            for (var j = 0; j < this.height; j++) {
+        for (var i = 0; i < self.width; i++) {
+            for (var j = 0; j < self.height; j++) {
                 var vertex = [i, j]
                 if (map[vertex] != 0 || vertex in done) continue